change dir name to plot_manager and refactor all change project name to plot_manager
