+++ README.md
@@ -1,2 +1,4 @@
-# contourPY
-contour map generation and animation in Python
+# plot_managerPY
+Matplotlib and seaborn plotting and analysis wrapper package in Python
+
+
