+++ plot_manager/__init__.py
@@ -1,0 +1,20 @@
+__version__ = "0.1.0"
+
+from plot_manager.plottype.DNA_track import DNAtrack
+from plot_manager.plottype.cluster_map import ClusterMap
+from plot_manager.plottype.contour import ContourMap
+from plot_manager.plottype.forest import Forest
+from plot_manager.plottype.image_stack import ImageStack
+from plot_manager.plottype.pie import Pie
+from plot_manager.plottype.protein_track import ProteinTrack
+from plot_manager.plottype.scatter import Scatter
+from plot_manager.plottype.survival import Survival
+from plot_manager.plottype.swarm_interval import SwarmInterval
+from plot_manager.plottype.track import Track
+from plot_manager.plottype.variant_track import VariantTrack
+from plot_manager.plottype.venn import Venn
+from plot_manager.plottype.violin import Violin
+from plot_manager.plottype.peristalsis import ZbPeristalsis
+from plot_manager.plottype.image import Image
+
+from plot_manager.plot_manager import PlotManager
