+++ plot_manager/plottype/ClusterMap.yml
