+++ plot_manager/plottype/DNA_track.py
@@ -2,7 +2,7 @@ import matplotlib.pyplot as plt
 import pandas as pd
 import numpy as np
 
-from plotmanager.plottype import track
+from plot_manager.plottype import track
 
 
 class DNAtrack(track.Track):
