+++ plot_manager/plottype/Image.yml
