+++ plot_manager/plottype/Scatter.yml
