+++ plot_manager/plottype/scatter.py
@@ -1,6 +1,6 @@
 import seaborn as sb
 
-from plotmanager.plottype import anim_plot
+from plot_manager.plottype import anim_plot
 
 
 class Scatter(anim_plot.AnimPlot):
