--- plotmanager/__init__.py
@@ -1,20 +1,0 @@
-__version__ = "0.1.0"
-
-from plotmanager.plottype.DNA_track import DNAtrack
-from plotmanager.plottype.cluster_map import ClusterMap
-from plotmanager.plottype.contour import ContourMap
-from plotmanager.plottype.forest import Forest
-from plotmanager.plottype.image_stack import ImageStack
-from plotmanager.plottype.pie import Pie
-from plotmanager.plottype.protein_track import ProteinTrack
-from plotmanager.plottype.scatter import Scatter
-from plotmanager.plottype.survival import Survival
-from plotmanager.plottype.swarm_interval import SwarmInterval
-from plotmanager.plottype.track import Track
-from plotmanager.plottype.variant_track import VariantTrack
-from plotmanager.plottype.venn import Venn
-from plotmanager.plottype.violin import Violin
-from plotmanager.plottype.peristalsis import ZbPeristalsis
-from plotmanager.plottype.image import Image
-
-from plotmanager.plot_manager import PlotManager
